Move generators functions outside of test cases
